+++ spotinst/ocean_aws_logging/consts.go
@@ -1,0 +1,11 @@
+package ocean_aws_logging
+
+import (
+	"github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
+)
+
+const (
+	Logging commons.FieldName = "logging"
+	Export  commons.FieldName = "export"
+	S3Id    commons.FieldName = "s3_id"
+)