+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* resource/spotinst_ocean_aws: added support for `logging`
+
 ## 1.63.0 (November 29, 2021)
 
 * resource/spotinst_ocean_gke_launch_spec: added support for `scheduling`